Presentation Slides
Originally, our colleagues had sourced incorrect  ROI for the S&P 500 and concluded that our trading bot outperformed the S&P 500. In the current ML_Trading_Bot python file and in the Read.me there is detailed explanation of the actual ROI for the S&P 500, which showed the S&P 500 out-performed the Apple stock machine learning bot.